Ignore 'require' function name appearing in resource parameters
The lint check previously checked all tokens and so erroneously detected
the 'require' metaparameter on a resource as a function call. All
matching tokens that are within a resource declaration are now excluded
from the lint check to ensure it only finds tokens in the main manifest
or class bodies.

Fixes #6